consolidate hook code and remove wrong template
We generally prefer to keep all custom hook code in a single file per
resource manager, so consolidated that into pkg/resource/key/hooks.go.
Also removed `templates/hooks/key/sdk_read_many_post_set_output.go.tpl`
because the Key resource uses a READ_ONE not READ_MANY getter call
(`DescribeKey` as opposed to `ListKeys`).
