+++ apis/v1alpha1/ack-generate-metadata.yaml
@@ -1,7 +1,7 @@
 ack_generate_info:
-  build_date: "2022-10-25T00:16:49Z"
+  build_date: "2022-11-08T14:11:18Z"
   build_hash: 5ee0ac052c54f008dff50f6f5ebb73f2cf3a0bd7
-  go_version: go1.19.2
+  go_version: go1.18.1
   version: v0.20.1-4-g5ee0ac0
 api_directory_checksum: 481291aa44a49278873c1188af55c812a4c18be6
 api_version: v1alpha1
