--- templates/hooks/key/sdk_read_many_post_set_output.go.tpl
@@ -1,3 +1,0 @@
-if err := rm.setResourceAdditionalFields(ctx, ko); err != nil {
-    return nil, err
-}
